requirements.txt
pandas==2.1.0
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.28.0
PyPDF2==3.0.1
pdfplumber==0.10.3
groq==0.4.0
python-dotenv==1.0.0
tiktoken==0.5.1
pytest==7.4.0
matplotlib==3.7.2
streamlit-option-menu==0.3.6